refactor: version 객체 생성 및 적용
