--- src/main/java/webserver/response/HttpVersion.java
@@ -1,16 +1,0 @@
-package webserver.response;
-
-public enum HttpVersion {
-    HTTP_1_1("HTTP/1.1"),
-    HTTP_2("HTTP/2");
-
-    private String version;
-
-    HttpVersion(String version) {
-        this.version = version;
-    }
-
-    public String getVersion() {
-        return version;
-    }
-}
